+++ src/Gate.IO.Api/Unified/Responses/GateUnifiedAccountInfo.cs
@@ -8,7 +8,7 @@ public record GateUnifiedAccountInfo
     /// <summary>
     /// Unified account mode
     /// </summary>
-    [JsonProperty("mode")]
+    [JsonProperty("mode"), JsonConverter(typeof(MapConverter))]
     public GateUnifiedAccountMode Mode { get; set; }
 
     /// <summary>